+++ apps/block_scout_web/assets/__mocks__/css/app.scss.js
@@ -1,4 +1,4 @@
 export default {
-  primary: "#a6ff33",
-  secondary: "#00bfff"
+  primary: "#496785",
+  secondary: "#00ffff"
 }