+++ CHANGELOG.md
@@ -999,7 +999,7 @@ fixed menu hovers in dark mode desktop view
 - [#2636](https://github.com/blockscout/blockscout/pull/2636) - Execute all address' transactions page queries in parallel
 - [#2596](https://github.com/blockscout/blockscout/pull/2596) - support AuRa's empty step reward type
 - [#2588](https://github.com/blockscout/blockscout/pull/2588) - add verification submission comment
-- [#2505](https://github.com/blockscout/blockscout/pull/2505) - support POA Network emission rewards
+- [#2505](https://github.com/blockscout/blockscout/pull/2505) - support Neatio Network emission rewards
 - [#2581](https://github.com/blockscout/blockscout/pull/2581) - Add generic Map-like Cache behaviour and implementation
 - [#2561](https://github.com/blockscout/blockscout/pull/2561) - Add token's type to the response of tokenlist method
 - [#2555](https://github.com/blockscout/blockscout/pull/2555) - find and show decoding candidates for logs